refactor: fold getProcessEnv into utils/env
